+++ src/main/java/com/github/mrglassdanny/mocalanguageserver/moca/lang/mocasql/util/MocaSqlLanguageUtils.java
@@ -3,9 +3,13 @@ package com.github.mrglassdanny.mocalanguageserver.moca.lang.mocasql.util;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import com.github.mrglassdanny.mocalanguageserver.moca.lang.MocaCompilationResult;
+import com.github.mrglassdanny.mocalanguageserver.moca.lang.mocasql.MocaSqlCompilationResult;
 import com.github.mrglassdanny.mocalanguageserver.moca.lang.mocasql.ast.MocaSqlSyntaxError;
 import com.github.mrglassdanny.mocalanguageserver.util.lsp.PositionUtils;
+import com.github.mrglassdanny.mocalanguageserver.util.lsp.RangeUtils;
 
+import org.antlr.v4.runtime.Token;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
 
@@ -67,7 +71,6 @@ public class MocaSqlLanguageUtils {
     }
 
     public static Range syntaxExceptionToRange(MocaSqlSyntaxError err, Range scriptRange) {
-
         return new Range(createMocaPosition(err.line, err.charPositionInLine, scriptRange),
                 createMocaPosition(err.line, err.charPositionInLine, scriptRange));
     }
@@ -82,4 +85,29 @@ public class MocaSqlLanguageUtils {
 
     }
 
+    public static Token getMocaSqlTokenAtPosition(Position pos, MocaCompilationResult mocaCompilationResult) {
+
+        for (MocaSqlCompilationResult mocaSqlCompilationResult : mocaCompilationResult.mocaSqlCompilationResults
+                .values()) {
+
+            if (RangeUtils.contains(mocaSqlCompilationResult.range, pos)) {
+                for (int i = 0; i < mocaSqlCompilationResult.mocaSqlTokens.size(); i++) {
+                    Token mocaSqlToken = mocaSqlCompilationResult.mocaSqlTokens.get(i);
+
+                    // Need to translate mocasqlToken positions into MOCA positions before we test
+                    // whether or not this is our token.
+                    Position adjMocasqlBeginTokenPos = createMocaPosition(mocaSqlToken.getLine(),
+                            mocaSqlToken.getCharPositionInLine(), mocaSqlCompilationResult.range);
+                    Position adjMocasqlEndTokenPos = new Position(adjMocasqlBeginTokenPos.getLine(),
+                            adjMocasqlBeginTokenPos.getCharacter() + mocaSqlToken.getText().length());
+
+                    if (RangeUtils.contains(new Range(adjMocasqlBeginTokenPos, adjMocasqlEndTokenPos), pos)) {
+                        return mocaSqlToken;
+                    }
+                }
+            }
+        }
+
+        return null;
+    }
 }